warehouse-project/requirements.txt
pandas==1.4.2
psycopg2==2.9.3
sqlalchemy==1.4.31
pyyaml==6.0
